requirements.txt
django
django-countries
djangorestframework
django-crispy-forms
django-import-export
xhtml2pdf
django-easy-pdf

